requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
absl-py==2.1.0
    # via mediapipe
aioice==0.9.0
    # via aiortc
aiortc==1.9.0
    # via streamlit-webrtc
altair==5.3.0
    # via streamlit
attrs==23.2.0
    # via
    #   jsonschema
    #   mediapipe
    #   referencing
av==12.2.0
    # via
    #   -r requirements.in
    #   aiortc
blinker==1.8.2
    # via
    #   flask
    #   streamlit
cachetools==5.3.3
    # via streamlit
certifi==2024.7.4
    # via requests
cffi==1.16.0
    # via
    #   aiortc
    #   cryptography
    #   pylibsrtp
    #   sounddevice
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   flask
    #   streamlit
colorama==0.4.6
    # via click
contourpy==1.2.1
    # via matplotlib
cryptography==42.0.8
    # via
    #   aiortc
    #   pyopenssl
cycler==0.12.1
    # via matplotlib
dnspython==2.6.1
    # via aioice
flask==3.0.3
    # via -r requirements.in
flatbuffers==24.3.25
    # via mediapipe
fonttools==4.53.1
    # via matplotlib
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via streamlit
google-crc32c==1.5.0
    # via aiortc
idna==3.7
    # via requests
ifaddr==0.2.0
    # via aioice
itsdangerous==2.2.0
    # via flask
jax==0.4.30
    # via mediapipe
jaxlib==0.4.30
    # via
    #   jax
    #   mediapipe
jinja2==3.1.4
    # via
    #   altair
    #   flask
    #   pydeck
joblib==1.4.2
    # via -r requirements.in
jsonschema==4.22.0
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
kiwisolver==1.4.5
    # via matplotlib
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
matplotlib==3.9.1
    # via mediapipe
mdurl==0.1.2
    # via markdown-it-py
mediapipe==0.10.14
    # via -r requirements.in
ml-dtypes==0.4.0
    # via
    #   jax
    #   jaxlib
numpy==2.0.0
    # via
    #   altair
    #   contourpy
    #   jax
    #   jaxlib
    #   matplotlib
    #   mediapipe
    #   ml-dtypes
    #   opencv-contrib-python
    #   opencv-python
    #   opt-einsum
    #   pandas
    #   pyarrow
    #   pydeck
    #   scipy
    #   streamlit
opencv-contrib-python==4.10.0.84
    # via mediapipe
opencv-python==4.10.0.84
    # via -r requirements.in
opt-einsum==3.3.0
    # via jax
packaging==24.1
    # via
    #   altair
    #   matplotlib
    #   streamlit
    #   streamlit-webrtc
pandas==2.2.2
    # via
    #   altair
    #   streamlit
pillow==10.4.0
    # via
    #   matplotlib
    #   streamlit
protobuf==4.25.3
    # via
    #   mediapipe
    #   streamlit
pyarrow==16.1.0
    # via streamlit
pycparser==2.22
    # via cffi
pydeck==0.9.1
    # via streamlit
pyee==11.1.0
    # via aiortc
pygments==2.18.0
    # via rich
pylibsrtp==0.10.0
    # via aiortc
pyopenssl==24.1.0
    # via aiortc
pyparsing==3.1.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2024.1
    # via pandas
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   -r requirements.in
    #   streamlit
rich==13.7.1
    # via streamlit
rpds-py==0.18.1
    # via
    #   jsonschema
    #   referencing
scipy==1.14.0
    # via
    #   jax
    #   jaxlib
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
sounddevice==0.4.7
    # via mediapipe
streamlit==1.36.0
    # via
    #   -r requirements.in
    #   streamlit-webrtc
streamlit-webrtc==0.47.7
    # via -r requirements.in
tenacity==8.5.0
    # via streamlit
toml==0.10.2
    # via streamlit
toolz==0.12.1
    # via altair
tornado==6.4.1
    # via streamlit
typing-extensions==4.12.2
    # via
    #   altair
    #   pyee
    #   streamlit
tzdata==2024.1
    # via pandas
urllib3==2.2.2
    # via requests
watchdog==4.0.1
    # via streamlit
werkzeug==3.0.3
    # via flask
